# Document/swagger-insure.v1.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "RFQ"
    },
    "servers": [
        {
            "url": "http://192.168.3.56:9763",
            "description": "Generated server url"
        }
    ],
    "paths": {
...
        "/rfq/policyTerms/updatePolicyTerms": {
            "put": {
                "tags": [
                    "policy-terms-controller"
                ],
                "operationId": "updatePolicyTerms",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PolicyTermsDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/rfq/policyTerms/createPolicyTerms": {
            "post": {
                "tags": [
                    "policy-terms-controller"
                ],
                "operationId": "createPolicyTerms",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PolicyTermsDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/policyType/addPolicyType/{id}": {
            "post": {
                "tags": [
                    "policy-type-controller"
                ],
                "operationId": "createPolicyType",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PolicyTypeDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PolicyTypeDto"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/rfq/policyTerms/getPolicyTermsById": {
            "get": {
                "tags": [
                    "policy-terms-controller"
                ],
                "operationId": "getPolicyTermsById",
                "parameters": [
                    {
                        "name": "rfqId",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PolicyTermsEntity"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/rfq/coverage/downloadMandateLetter/{rfqId}": {
            "get": {
                "tags": [
                    "coverage-details-controller"
                ],
                "operationId": "downloadMandateLetter",
                "parameters": [
                    {
                        "name": "rfqId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string",
                                        "format": "byte"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/policyType/getPolicyTypeById/{id}": {
            "get": {
                "tags": [
                    "policy-type-controller"
                ],
                "operationId": "getPolicyTypeById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/rfq/policyTerms/deletePolicyTermsById/{rfqId}": {
            "delete": {
                "tags": [
                    "policy-terms-controller"
                ],
                "operationId": "deletePolicyTermsById",
                "parameters": [
                    {
                        "name": "rfqId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "PolicyTermsChildDto": {
                "type": "object",
                "properties": {
                    "policyTermId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "coverageName": {
                        "type": "string"
                    },
                    "remark": {
                        "type": "string"
                    }
                }
            },
            "PolicyTermsDto": {
                "type": "object",
                "properties": {
                    "rfqId": {
                        "type": "string"
                    },
                    "policyDetails": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PolicyTermsChildDto"
                        }
                    },
                    "createDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updateDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "recordStatus": {
                        "type": "string"
                    }
                }
            },
...
            "PolicyTypeDto": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "name": {
                        "type": "string"
                    },
                    "policyId": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
...
            "PolicyTermsEntity": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "rfqId": {
                        "type": "string"
                    },
                    "coverageName": {
                        "type": "string"
                    },
                    "remark": {
                        "type": "string"
                    },
                    "createDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updateDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "recordStatus": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}